# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/prompts": {
      "get": {
        "operationId": "AdminPromptsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "staff": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminPromptsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromptDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "staff": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/prompts/{id}": {
      "patch": {
        "operationId": "AdminPromptsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePromptDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "staff": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/calendar": {
      "get": {
        "operationId": "AdminCalendarController_list",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "staff": []
          }
        ],
        "summary": "Grade de publicação (dia × slot × trilha × idioma) com estados",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Você conhece Deus? — API",
    "description": "API do app e do painel administrativo. Este documento é a fonte de verdade do contrato: gera o client Dart (mobile) e os tipos TS (admin) em packages/contracts.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "user": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "staff": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePromptDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePromptDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}